+++ database/nostr-ndb/src/lib.rs
@@ -77,6 +77,13 @@ impl NostrDatabase for NdbDatabase {
         Backend::LMDB
     }
 
+    fn features(&self) -> Features {
+        Features {
+            persistent: true,
+            full_text_search: true,
+        }
+    }
+
     fn save_event<'a>(
         &'a self,
         event: &'a Event,